+++ modules/item/cpr-item.js
@@ -18,6 +18,7 @@ export default class CPRItem extends Item {
   /* -------------------------------------------- */
   /** @override */
   prepareData() {
+    // eslint-disable-next-line foundry-cpr/logger-after-function-definition
     LOGGER.trace(`prepareData | CPRItem | Called for type: ${this.type}.`);
     super.prepareData();
   }
@@ -71,7 +72,7 @@ export default class CPRItem extends Item {
   // Any calls to functions not related to rolls, triggered from actions.
   // actorSheet UX gets actived -> actorSheet.eventFunction(event) ->
   doAction(actor, actionAttributes) {
-    LOGGER.debug("doAction | CPRItem | Called.");
+    LOGGER.trace("doAction | CPRItem | Called.");
     const itemType = this.data.type;
     // const changedItems = [];
     switch (itemType) {
@@ -173,7 +174,7 @@ export default class CPRItem extends Item {
   // ammo Item Methods
   // TODO - REFACTOR, do not do this...
   _ammoAction(actionAttributes) {
-    LOGGER.debug("_ammoAction | CPRItem | Called.");
+    LOGGER.trace("_ammoAction | CPRItem | Called.");
     const actionData = actionAttributes["data-action"].nodeValue;
     const ammoAmount = actionAttributes["data-amount"].nodeValue;
     switch (actionData) {
@@ -194,14 +195,14 @@ export default class CPRItem extends Item {
 
   // SKILL FUNCTIONS
   setSkillLevel(value) {
-    LOGGER.debug("setSkillLevel | CPRItem | Called.");
+    LOGGER.trace("setSkillLevel | CPRItem | Called.");
     if (this.type === "skill") {
       this.getData().level = Math.clamped(-99, value, 99);
     }
   }
 
   setSkillMod(value) {
-    LOGGER.debug("setSkillMod | CPRItem | Called.");
+    LOGGER.trace("setSkillMod | CPRItem | Called.");
     if (this.type === "skill") {
       this.getData().skillmod = Math.clamped(-99, value, 99);
     }
@@ -218,7 +219,7 @@ export default class CPRItem extends Item {
 
   // AMMO FUNCTIONS
   async _ammoDecrement(changeAmount) {
-    LOGGER.debug("_ammoDecrement | CPRItem | Called.");
+    LOGGER.trace("_ammoDecrement | CPRItem | Called.");
     const currentValue = this.data.data.amount;
     const newValue = Math.max(0, Number(currentValue) - Number(changeAmount));
     this.data.data.amount = newValue;
@@ -228,7 +229,7 @@ export default class CPRItem extends Item {
   }
 
   async _ammoIncrement(changeAmount) {
-    LOGGER.debug("_ammoIncrement | CPRItem | Called.");
+    LOGGER.trace("_ammoIncrement | CPRItem | Called.");
     const currentValue = this.data.data.amount;
     const newValue = Number(currentValue) + Number(changeAmount);
     this.data.data.amount = newValue;
@@ -240,7 +241,7 @@ export default class CPRItem extends Item {
   // Weapon Item Methods
   // TODO - Refactor
   async _weaponAction(actor, actionAttributes) {
-    LOGGER.debug("_weaponAction | CPRItem | Called.");
+    LOGGER.trace("_weaponAction | CPRItem | Called.");
     const actionData = actionAttributes["data-action"].nodeValue;
     switch (actionData) {
       case "select-ammo":
@@ -266,7 +267,7 @@ export default class CPRItem extends Item {
   }
 
   async _measureDv(actor, dvTable) {
-    LOGGER.debug("_measureDv | CPRItem | Called.");
+    LOGGER.trace("_measureDv | CPRItem | Called.");
     if (actor.sheet.token !== null) {
       actor.sheet.token.update({ "flags.cprDvTable": dvTable });
     }
@@ -274,7 +275,7 @@ export default class CPRItem extends Item {
 
   // TODO - Refactor
   async _weaponUnload() {
-    LOGGER.debug("_weaponUnload | CPRItem | Called.");
+    LOGGER.trace("_weaponUnload | CPRItem | Called.");
     if (this.actor) {
       // recover the ammo to the right object
       const { ammoId } = this.data.data.magazine;
@@ -299,7 +300,6 @@ export default class CPRItem extends Item {
   async _weaponLoad(reloadAmmoId) {
     LOGGER.trace("_weaponLoad | CPRItem | Called.");
     let selectedAmmoId = reloadAmmoId;
-    LOGGER.debug("_weaponLoad | CPRItem | Called.");
     const loadUpdate = [];
     if (this.actor) {
       if (!selectedAmmoId) {
@@ -394,7 +394,7 @@ export default class CPRItem extends Item {
   }
 
   setWeaponAmmo(value) {
-    LOGGER.debug("setWeaponAmmo | CPRItem | Called.");
+    LOGGER.trace("setWeaponAmmo | CPRItem | Called.");
     const maxAmmo = this.getData().magazine.max;
     if (this.type === "weapon") {
       if (value.charAt(0) === "+" || value.charAt(0) === "-") {
@@ -406,7 +406,7 @@ export default class CPRItem extends Item {
   }
 
   setItemAmount(value) {
-    LOGGER.debug("setItemAmount | CPRItem | Called.");
+    LOGGER.trace("setItemAmount | CPRItem | Called.");
     if (value.charAt(0) === "+" || value.charAt(0) === "-") {
       this.getData().amount = this.getData().amount + parseInt(value, 10);
     } else {
@@ -714,12 +714,12 @@ export default class CPRItem extends Item {
     const { weaponType } = this.data.data;
     let { damage } = this.data.data;
     let universalBonusDamage = this.actor.data.data.universalBonuses.damage;
-    if (weaponType === "unarmed" && this.data.data.unarmedAutomaticCalculation) {
+    if ((weaponType === "unarmed" || weaponType === "martialArts") && this.data.data.unarmedAutomaticCalculation) {
       // calculate damage based on BODY stat
       const actorBodyStat = this.actor.data.data.stats.body.value;
       if (actorBodyStat <= 4) {
-        if (this.actor.data.filteredItems.cyberware.some((c) => ((c.data.data.type === "cyberArm") && (c.data.data.isInstalled === true) && (c.data.data.isFoundational === true)))) {
-          // If the user has an installed Cyberarm, which is a foundational
+        if (weaponType === "unarmed" && this.actor.data.filteredItems.cyberware.some((c) => ((c.data.data.type === "cyberArm") && (c.data.data.isInstalled === true) && (c.data.data.isFoundational === true)))) {
+          // If the user has an installed Cyberarm, which is a foundational. This is only for unarmed damage, not martial arts damage.
           damage = "2d6";
         } else {
           damage = "1d6";
@@ -833,7 +833,7 @@ export default class CPRItem extends Item {
    * @public
    */
   setInstalled() {
-    LOGGER.debug("setInstalled | CPRItem | Called.");
+    LOGGER.trace("setInstalled | CPRItem | Called.");
     if (this.data.type !== "program") {
       return;
     }
@@ -846,7 +846,7 @@ export default class CPRItem extends Item {
    * @public
    */
   unsetInstalled() {
-    LOGGER.debug("unsetInstalled | CPRItem | Called.");
+    LOGGER.trace("unsetInstalled | CPRItem | Called.");
     if (this.data.type !== "program") {
       return;
     }
@@ -859,7 +859,7 @@ export default class CPRItem extends Item {
    * @public
    */
   getInstalled() {
-    LOGGER.debug("getInstalled | CPRItem | Called.");
+    LOGGER.trace("getInstalled | CPRItem | Called.");
     if (this.data.type !== "program") {
       return;
     }
@@ -880,7 +880,7 @@ export default class CPRItem extends Item {
    * @public
    */
   availableSlots() {
-    LOGGER.debug("availableSlots | CPRItem | Called.");
+    LOGGER.trace("availableSlots | CPRItem | Called.");
     const itemData = duplicate(this.data.data);
 
     let unusedSlots = 0;
@@ -927,7 +927,7 @@ export default class CPRItem extends Item {
    * @public
    */
   getInstalledPrograms() {
-    LOGGER.debug("getInstalledPrograms | CPRItem | Called.");
+    LOGGER.trace("getInstalledPrograms | CPRItem | Called.");
     return this.data.data.programs.installed;
   }
 
@@ -940,7 +940,7 @@ export default class CPRItem extends Item {
    * @public
    */
   getRezzedPrograms() {
-    LOGGER.debug("getRezzedPrograms | CPRItem | Called.");
+    LOGGER.trace("getRezzedPrograms | CPRItem | Called.");
     return this.data.data.programs.rezzed;
   }
 
@@ -951,7 +951,7 @@ export default class CPRItem extends Item {
    * @param {Array} programs      - Array of CPRItem programs
    */
   installPrograms(programs) {
-    LOGGER.debug("installPrograms | CPRItem | Called.");
+    LOGGER.trace("installPrograms | CPRItem | Called.");
     const { installed } = this.data.data.programs;
     programs.forEach((p) => {
       const onDeck = installed.filter((iProgram) => iProgram._id === p.data._id);
@@ -972,7 +972,7 @@ export default class CPRItem extends Item {
    * @param {Array} programs      - Array of CPRItem programs
    */
   uninstallPrograms(programs) {
-    LOGGER.debug("uninstallPrograms | CPRItem | Called.");
+    LOGGER.trace("uninstallPrograms | CPRItem | Called.");
     let { rezzed } = this.data.data.programs;
     let { installed } = this.data.data.programs;
     const tokenList = [];
@@ -1014,7 +1014,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program to check
    */
   isRezzed(program) {
-    LOGGER.debug("isRezzed | CPRItem | Called.");
+    LOGGER.trace("isRezzed | CPRItem | Called.");
     const rezzedPrograms = this.data.data.programs.rezzed.filter((p) => p._id === program.id);
     const { installed } = this.data.data.programs;
     const installIndex = installed.findIndex((p) => p._id === program.data._id);
@@ -1036,7 +1036,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program to REZ
    */
   async rezProgram(program, callingToken) {
-    LOGGER.debug("rezProgram | CPRItem | Called.");
+    LOGGER.trace("rezProgram | CPRItem | Called.");
     const programData = duplicate(program.data);
     const { installed } = this.data.data.programs;
     const installIndex = installed.findIndex((p) => p._id === programData._id);
@@ -1069,7 +1069,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program create the Token for
    */
   async _rezBlackIceToken(programData, callingToken) {
-    LOGGER.debug("_rezBlackIceToken | CPRItem | Called.");
+    LOGGER.trace("_rezBlackIceToken | CPRItem | Called.");
     let netrunnerToken = callingToken;
     let scene;
     const blackIceName = programData.name;
@@ -1168,7 +1168,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program de-rez
    */
   async derezProgram(program) {
-    LOGGER.debug("derezProgram | CPRItem | Called.");
+    LOGGER.trace("derezProgram | CPRItem | Called.");
     const { installed } = this.data.data.programs;
     const installIndex = installed.findIndex((p) => p._id === program.id);
     const programState = installed[installIndex];
@@ -1191,7 +1191,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program to remove the token for
    */
   async _derezBlackIceToken(programData) {
-    LOGGER.debug("_derezBlackIceToken | CPRItem | Called.");
+    LOGGER.trace("_derezBlackIceToken | CPRItem | Called.");
     if (typeof programData.flags["cyberpunk-red-core"] !== "undefined") {
       const cprFlags = programData.flags["cyberpunk-red-core"];
       const { biTokenId } = cprFlags;
@@ -1224,7 +1224,7 @@ export default class CPRItem extends Item {
    * @param {CPRItem} program      - CPRItem of the program to reset
    */
   resetRezProgram(program) {
-    LOGGER.debug("resetRezProgram | CPRItem | Called.");
+    LOGGER.trace("resetRezProgram | CPRItem | Called.");
     const { rezzed } = this.data.data.programs;
     const rezzedIndex = rezzed.findIndex((p) => p._id === program.id);
     const { installed } = this.data.data.programs;
@@ -1240,7 +1240,7 @@ export default class CPRItem extends Item {
    * @param {Number} reduceAmount - Amount to reduce REZ by. Defaults to 1.
    */
   reduceRezProgram(program, reduceAmount = 1) {
-    LOGGER.debug("reduceRezProgram | CPRItem | Called.");
+    LOGGER.trace("reduceRezProgram | CPRItem | Called.");
     const { rezzed } = this.data.data.programs;
     const rezzedIndex = rezzed.findIndex((p) => p._id === program.id);
     const programState = rezzed[rezzedIndex];
@@ -1261,7 +1261,7 @@ export default class CPRItem extends Item {
   }
 
   updateRezzedProgram(programId, updatedData) {
-    LOGGER.debug("updateRezzedProgram | CPRItem | Called.");
+    LOGGER.trace("updateRezzedProgram | CPRItem | Called.");
     const { rezzed } = this.data.data.programs;
     const rezzedIndex = rezzed.findIndex((p) => p._id === programId);
     const programState = rezzed[rezzedIndex];
@@ -1291,7 +1291,7 @@ export default class CPRItem extends Item {
    * @param {String} boosterType - string defining the type of boosters to return.
    */
   getBoosters(boosterType) {
-    LOGGER.debug("getBoosters | CPRItem | Called.");
+    LOGGER.trace("getBoosters | CPRItem | Called.");
     const { rezzed } = this.data.data.programs;
     let modifierTotal = 0;
     switch (boosterType) {
@@ -1323,7 +1323,8 @@ export default class CPRItem extends Item {
   }
 
   _createCyberdeckRoll(rollType, actor, extraData = {}) {
-    LOGGER.debug("_createCyberdeckRoll | CPRItem | Called.");
+    LOGGER.trace("_createCyberdeckRoll | CPRItem | Called.");
+    const rollTitle = "";
     let rollModifiers = 0;
     let cprRoll;
     const { programId } = extraData;